logically separated hardware from fri ros
