+++ lbr_fri_ros2/include/lbr_fri_ros2/lbr_app.hpp
@@ -10,7 +10,7 @@
 #include "std_srvs/srv/empty.hpp"
 
 #include "lbr_fri_msgs/srv/app_connect.hpp"
-#include "lbr_hardware_new/lbr_state_client.hpp"
+#include "lbr_fri_ros2/lbr_state_client.hpp"
 
 #include "fri/friUdpConnection.h"
 #include "fri/friClientApplication.h"
