+++ lbr_fri_ros2/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>lbr_hardware_new</name>
+  <name>lbr_fri_ros2</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
   <maintainer email="martin.huber@kcl.ac.uk">mhubii</maintainer>
