+++ lbr_fri_ros2/src/lbr_app.cpp
@@ -1,4 +1,4 @@
-#include "lbr_hardware_new/lbr_app.hpp"
+#include "lbr_fri_ros2/lbr_app.hpp"
 
 namespace lbr_hardware
 {
